move test to be under launch
