+++ tests/pytest_tests/system_tests/test_core/test_footer.py
@@ -1,7 +1,5 @@
 """footer tests."""
 
-import re
-
 import numpy as np
 import pytest
 import wandb
@@ -121,23 +119,3 @@ def test_footer_history(wandb_init, check_output_fn):
     run.log(dict(a=3))
     run.finish()
     check_output_fn(exp_summary=[], exp_history=["a", "d", "🚀"])
-
-
-# todo(core): implement job info
-@pytest.mark.wandb_core_failure(feature="launch")
-def test_footer_job_output(wandb_init, capsys, monkeypatch):
-    """Test that footer includes job info when a job is created."""
-    monkeypatch.setenv("WANDB_DOCKER", "hello-world")  # Needed to trigger job creation.
-    run = wandb_init()
-    run.log({"a": 1})
-    run.finish()
-    captured = capsys.readouterr()
-    lines = captured.err.splitlines()
-    for line in lines:
-        if re.search(
-            r"View job at http://localhost:8080/user-\w+-\w+/uncategorized/jobs/[\w=]+/version_details/v0",
-            line,
-        ):
-            break
-    else:
-        raise AssertionError(f"Job URL not found in footer: {captured.err}")
