+++ tests/pytest_tests/system_tests/test_launch/test_launch_run.py
@@ -1,6 +1,8 @@
 import json
+import re
 from unittest import mock
 
+import pytest
 from wandb.sdk.internal import job_builder
 
 
@@ -30,3 +32,22 @@ def test_run_use_job_env_var(runner, relay_server, test_settings, user, wandb_in
                 if "mutation UseArtifact(" in data.get("request", {}).get("query", ""):
                     use_count += 1
         assert use_count == 1
+
+
+@pytest.mark.wandb_core_failure(feature="launch")
+def test_footer_job_output(wandb_init, capsys, monkeypatch):
+    """Test that footer includes job info when a job is created."""
+    monkeypatch.setenv("WANDB_DOCKER", "hello-world")  # Needed to trigger job creation.
+    run = wandb_init()
+    run.log({"a": 1})
+    run.finish()
+    captured = capsys.readouterr()
+    lines = captured.err.splitlines()
+    for line in lines:
+        if re.search(
+            r"View job at http://localhost:8080/user-\w+-\w+/uncategorized/jobs/[\w=]+/version_details/v0",
+            line,
+        ):
+            break
+    else:
+        raise AssertionError(f"Job URL not found in footer: {captured.err}")
